functions/external-model-proxy/requirements.txt
flask
azure-identity
gunicorn
